requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.1
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
interchange==2021.0.4
monotonic==1.6
numpy==1.26.4
openai==1.35.6
packaging==24.1
pansi==2020.7.3
py2neo==2021.2.4
pydantic==2.7.4
pydantic-core==2.18.4
pygments==2.18.0
python-dotenv==1.0.1
pytz==2024.1
pyyaml==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
six==1.16.0
sniffio==1.3.1
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.42.1
typing-extensions==4.12.2
urllib3==2.2.2
